requirements.txt
asgiref==3.3.1
Django==3.1.4
django-taggit==1.3.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
uuid==1.30
uWSGI==2.0.18
whitenoise==5.2.0
